Bug-fixing unit changes, improved precision
Fixed some bugs that made the program crash from  both integer overflows and underflows, capacity overflows and such. Furthermore, the 'average' field in Magnitude is now a float, allowing for more precision, specially notable in the GPU util. Also, network speed now uses Bytes per second cause I like it better (should probably add a custom units thingie down the line??).

+++ src/cpu.rs
@@ -73,8 +73,8 @@ impl Cpu {
                     util,
                     all.split(' ')
                         .skip(1)
-                        .map(|s| s.parse::<u32>().expect("fully numeric string"))
-                        .collect::<Vec<u32>>(),
+                        .map(|s| s.parse::<usize>().expect("fully numeric string"))
+                        .collect::<Vec<usize>>(),
                 )
             })
             .for_each(|(total, idle, util, all)| {
